* python/tag.cc: Support 'key in mapping' for TagSections
Support the replacement of mapping.has_key() for sections, and
update the usage in apt/package.py and apt/debfile accordingly.

This is implemented by extending the TagSecType with sequence
methods, but only settings the contains method there.

The TagSecGetAttr() function has been removed and replaced by
the use of the tp_methods slot.