* apt/*.py: Initial rename work for Bug#481061
A new module, apt.deprecation, is introduced containing functions and
classes which assist in the deprecation.

The apt_pkg extension gets a new attribute, _COMPAT_0_7 which can be set
by defining COMPAT_0_7 at compile time (-DCOMPAT_0_7).
      
The names are changed, and compatibility functions are enabled if
bool(apt_pkg._COMPAT_0_7) == True, i.e. if the package has been built with
backward compatibility fixes.

This commit changes the apt and aptsources packages, the apt_pkg and
apt_inst extensions will be the next renames.