* utils/migrate-0.8.py: Correctly import modules, improve attribute detection
In order to import modules from a package, 'fromlist' may not be empty. Therefore,
we pass ['*'] now.

When attributes where checked, we just checked their names and did not check
their classes. This meant that in e.g.
	Compat-API: A.alpha, B.alpha
	Clean-API:  A.alpha
The attribute 'alpha' would not be considered deprecated because it is provided
by A.alpha. Now we treat an 'alpha' attribute as deprecated, if at least one class
loses it.