* debian/rules: Support multiple PY3K versions.
Introduce PY3K_VERSIONS, a list of all py3k versions the package should
be built for. The default is to build for the currently installed versions.

Introduce 2TO3_VERSION to select a 2to3 version which will convert the
python code to python 3. Points to the latest version by default.