python/progress.cc, apt/progress/*.py: Use PEP 8 naming conventions for progress
This changes the progress classes to use PEP 8 names. Due to the concept of the
deprecation system, this causes methods of subclasses not to be called at all.

If a class implements a command with underscores, it is called. If the class
also implements the version in mixedCase, this is ignored. This means that
all subclasses will not work correctly, because only the method from the
parent class is called.