Bugfix: Delete pointers correctly, fixing memory leaks. (LP: #370149)
We previously called the destructor of the pointer. This resulted in no
object using pointers being deallocated.

This patch introduces CppDeallocPtr() and CppOwnedDeallocPtr() which do
the same as the other CppDealloc() and CppOwnedDealloc(), but use 'delete'
on the pointer instead of the deconstructor.

Furthermore, this patch also changes AcquireFile to be a CppOwnedPyObject,
owned by the Acquire object. Without this change, deleting the Acquire
object would cause a crash when AcquireFile is deallocated.