python/acquire.cc: Make AcquireFile a subclass of AcquireItem
Generalized the code a bit, so we can now access the various attributes
of AcquireItem in the AcquireFile (pkgAcqFile is a subclass of
pkgAcquire::Item).

This will allow us to implement a raw object with a single pointer to
an Item later, which we will need for the new progress interface.