python/generic.h: Introduce NoDelete field in CppPyObject.
Setting NoDelete to true causes the deallocation functions to not delete
the underlying C++ object. This is useful in situations where the object
is already managed somewhere else, e.g. DepCache by CacheFile.

It will also assist us in providing a flexible C++ API which only takes
ownership if requested.