python/acquire.cc: Fix segmentation faults, introduce PyAcquireObject.
Make AcquireItem objects raise ValueError instead of segfaulting when the Acquire()
object is shut down or the main object (e.g. AcquireFile) is deallocated.

This is implemented by using a vector of the AcquireItem objects, and setting
AcquireItem->Object = NULL, when the memory 'Object' previously pointed to is
going to be deleted.