apt/debfile.py: Adapt to class-based API (WARNING: changes behavior on certain invalid packages).
Use the class-based API now. This also means that if the archive is no
valid Debian package (because it misses a 'data.tar.*', 'control.tar.gz'
or a 'debian-binary' member) the method open() will now raise an Error;
previously it only raised an error if there was no 'debian-binary'
member or no 'control.tar.gz' member.