apt/cache.py: Support :all and other special architecture specifies
Remove the _set and _fullnameset that previously stored the names of
all packages in the cache.

Checking if a package is now a matter of checking whether it is
in the low-level cache and has versions (that is, is real).

Still keep the _sorted_set around and make keys() manage it, and
change everyone using _sorted_set for iteration to use keys().