apt.Version: Compare against package name and version (if available)
Before this commit the apt.Version._cmp method only compared the version
string.  The package name was not part of the comparison and hence apt.Version
objects with the same version but different package names would have been
falsely equal.

With this commit it is still possible to compare an apt.Version object against
a version string for backwards compatibility.