arfile, tarfile: Decode member names using filesystem encoding
It would be nice if we also supported the use case where, if you
open the archive file using a bytes path, you also get back bytes
members, but that seems optional.

The PyUnicode_DecodeFSDefault() function uses the surrogateescape
error handler, so it should handle differently encoded paths as
well.

Closes: #806596