Adjust dependencies for apt.auth on gnupg and dirmngr
We need dirmngr for gnupg (>> 2) [maybe 2.1 formally, but who
cares] in order for keyserver updates to work. We also need
gnupg in any case, as apt.auth calls apt-key stuff that
depends on it.

Note that this should be dropped to a recommends, and apt.auth
should be properly deprecated.