Work around broken debfile multiarch test cases
Not sure why the first one is broken now, but let's add a check()
call there as the message tells us. The second one is obvious, but
I do not want to fix it correctly -- the overhead is just too much
right now (proper fake chroot with fake packages, no infra for that
yet).