Debian template: Adopt MatchUri to match base, switch BaseUri to CDN
The BaseUri was defined as http.us.debian.org, but that does not
match the MatchUri field, causing no source to be selected. Change
the MatchUri field to also accept mirrors with http instead of ftp
in their name.

Also switch the default BaseUri to deb.debian.org, and adapt the
MatchUri to understand that.

Closes: #859672