DepCache: Check that candidate we are setting belongs to package
We expose an API that takes a Package and a Version object,
but the APT API only takes a version, and therefore we are
not checking the package at all.

This meant, for example, that, in the high-level "apt" module,
where the candidate is set via a property, something like that:

	cache["apt"].candidate = cache["dpkg"].versions[<something>]

would have set the candidate for dpkg, and not for apt.