Fix InstallProgress for installing .deb files on Python >= 3.4
Since PEP-446 [1] got implemented in Python 3.4, all file descriptors
are created CLOEXEC, but we need to be able to pass the writefd of our
pipe if we spawn dpkg for installing a single .deb file.

This means that currently in this situation dpkg is getting an invalid
file descriptor and aborts right away.

Fix this by marking the file descriptor as inheritable by child
processes if the API introduced by the conflicting PEP-446 is
available, which also means we're in a situation that needs fixing.

[1] https://www.python.org/dev/peps/pep-0446/