distinfo: correct subprocess usage
The release codename is retrieved using a call to lsb_release, but the
result is bytes, not a str, so later usage is mixed into a string path
and fails to correctly find the properly named template file.

The call can throw IOError in python3 (actually FileNotFoundError,
but that error doesn't exist in python2) so catch that also.