aptsources: Allow non-existent `source.list` file
Apt itself allows for the `sources.list` file to not exist,
and so should aptsources.

For example, the `puppetlabs/apt` Puppet module changed its
purging behavior for `sources.list` file to simply delete
the file instead of leaving a file with a comment as it
previously did. As a result, everyone enforcing source entry
purging via the Puppet module as a result receives an unnecessary
warning when this Python module is used prior to this change.