Use pybuild to split out -dbg stuff, avoid .install/.files
Doing this via pybuild is somewhat nasty, as it does not
allow us to move all the common files into the common
package itself, so we have to hack around that.

I'd like to more declaratively state which files belong
into common, or for pybuild to figure that out itself
(it ain't hard), but meh, what can I do?

This fixes the problem where the wildcard that was
matching debug symbols "d-*" (formerly dm-*) also is
matching kfreebsd-, and hence caused FTBFS on freebsd.

Closes: #958118