Do not require hardcoding release names in distro info, cleanup Debian and Ubuntu
Read the data from distro-info-data and expand templates
in-memory such that we don't have to duplicate data.

For Ubuntu, this changes some historical releases a bit more, but the
changes do seem correct, and the test suite still passes. Mostly just
some reorganization that was done in later releases passed down to older
releases.

LP: #1727470